requirements.txt
fastapi
uvicorn
pypdf
requests
google-generativeai
chromadb
sentence-transformers
nest-asyncio
pyngrok
